Clean up the gitlab CI based on newly open sourced model components (#19)
This PR removes more dead code as we transition to open source.
Specifically, we moved from renovate to dependabot and no longer need to
clone `firewheel_repo_base` or `firewheel_repo_tutorials` as they are
now open source. Lastly, we no longer need to have a deploy function as
that was replaced by the corresponding GitHub action.