feat: resolve #33 by enhancing the 'dry-run' flag (#50)
This PR resolves #33 by updating the `dry-run` flag for the `experiment`
helper to actually import plugins and model component objects. This will
then enable us to catch syntax errors much earlier in the development
process. While this may not entirely resolve every use-case for #33, it
does cover the intent and a new issue can always be opened for whatever
specific use case is.

This PR passes the dry-run flag to the `process_model_component`
function and ensures that importing the plugin and model component
objects is tried regardless of whether or not a dry-run is occurring.
This enables catching SyntaxErrors, ImportErrors, IndentationErrors,
among others all without executing the plugin's `run` function. This has
the added bonus of simplifying the logic for the `experiment` helper.

---------

Signed-off-by: Steven Elliott <6461548+sdelliot@users.noreply.github.com>
Co-authored-by: Mitch Negus <21086604+mitchnegus@users.noreply.github.com>