Update documentation with minor changes (#56)
This provides a few minor updates to the docs.

1. I clarify that the `PYTHON_BIN` environment variable is the
executable (which may be either a path or a symlink). I think this
wording might be better, because just giving `python3` (assuming that
it's being run in a virtual environment) doesn't strike me as a path.
2. It appears that `libenchant-dev` has been superseded by
`libenchant-2-dev` and is accessible with that new name in the relevant
Ubuntu repositories. (Similar for Fedora-based OSs and `enchant2-devel`
and `enchant-devel`.)