ci: Add new CI actions to autotag PRs and update release notes and changelog entries (#62)
This PR is an effort to improve the OSS development process for
FIREWHEEL. In it, we add the following GitHub actions:

-
[amannn/action-semantic-pull-request](https://github.com/amannn/action-semantic-pull-request)
- Validate that PR titles have the general style of [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/). We have also
extended the default list to add a few new commit "types".
-
[release-drafter/release-drafter](https://github.com/release-drafter/release-drafter)
- Drafts the next release notes as pull requests are merged into main.
- This action also includes an "autolabeler" which checks the branch
name and PR title to automatically add a label to the request.
-
[danielchabr/pr-labels-checker](https://github.com/danielchabr/pr-labels-checker)
- Github Action to check if a PR has at least one of the expected labels
-
[stefanzweifel/changelog-updater-action](https://github.com/stefanzweifel/changelog-updater-action)
- A GitHub Action to automatically update a "Keep a Changelog" CHANGELOG
with the latest release notes.

These four actions work together to ensure that each PR will be
accurately labeled and that, when merged, will ensure that the release
notes are correctly updated. Finally, when we issue a release, the
changelog will also be updated. This has been tested in a separate
repository and seems to work as expected.

Note that this PR will fail the CI pipeline as the release-drafter
configuration is not in our default branch yet.