refactor: Replace the gRPC-backed RepositoryDB with a JSON-file backed one. (#48)
This is an effort to reduce the impact of our services on starting and
using FIREWHEEL. Specifically, the RepositoryDB has been part of the
gRPC service that FIREWHEEL provides. This inter-dependency is
unnecessary for the following reasons:

1. It does not need to be accessed except from the "control node".
2. The gRPC service stored the repositories (e.g., a path) in a JSON
file.
3. As repositories become pip packages, the necessity of this feature
wains entirely.

In this PR, we strip away the unnecessary complexity of gRPC and simply
rely on the single JSON file for locally-installed model components. We
have updated all necessary documentation and tests accordingly. Finally,
we needed to provide a minimum version for the `grpcio` package as the
newly generated code checks version information.

Note that the `pb2*` files are auto-generated via `grpc_tools.protoc`.

---------

Signed-off-by: Steven Elliott <6461548+sdelliot@users.noreply.github.com>
Co-authored-by: Mitch Negus <21086604+mitchnegus@users.noreply.github.com>