feat: Create a initial docker container for FIREWHEEL (#68)
This PR provides a working Docker container for FIREWHEEL. We should
consider this as a beta-testing phase and remove that label in a future
release. This container has been tested in both low-trust and high-trust
environments where various permissions/kernel modules were stripped and
verified functionality. The associated documentation provides a good
overview of the container and how it was created.
Aside from that, there are a few important things to note:
- The FIREWHEEL docker container is based on the
[minimega](https://github.com/sandia-minimega/minimega/) container
image.
- Multi-node/containers have not been tested/explored.
- This PR does not include a Docker compose environment. This should
happen in a different PR.
- The container does not contain any VM images and those will need to be
added post-launch by users.

---------

Signed-off-by: Steven Elliott <6461548+sdelliot@users.noreply.github.com>