ci: Fix the testing, linting, and documentation CI workflows (#95)
Remote references in the GitHub workflows to composite actions must
include an `@ref` tag or they won't be valid (the current state of
things).

This adds a `@v2` ref to each of the composite actions called by the
FIREWHEEL CI.

## Important Notes

Changes to the actions in this repo will **not** be automatically tested
on their PRs because the actions called during the CI pipeline will be
run from the latest tagged version. A solution to this _might_* be to
test changes to those composite actions by also adjusting the workflows
calling those actions in this repository to use local references to the
actions.

<small>*Full disclosure: I have not tested this approach.</small>

We expect that changes to the composite actions here will generally be
infrequent enough that this approach will be tolerable.

We have made the decision to prioritize using the full path
(`owner/repo@ref`) to the action rather than the local path so that this
repository's actions can be shared among other tools in the FIREWHEEL
ecosystem. As an example, `firewheel_repo_base` performs linting by
referencing the reusable linting workflow in this repository. However,
when that reusable workflow is run in the `firewheel_repo_base` CI
pipeline, it looks for any actions defined locally in the `firewheel`
repo in the `firewheel_repo_base` repo—obviously it does not find them
there. Due to external considerations, we have chosen not to create
individual composite action repositories as suggested by GitHub.
Instead, by specifying the absolute remote path, we set FIREWHEEL model
components to go directly to the `firewheel` repo for all of these
composite actions.

Breaking changes to the installation process (the majority of the
composite actions currently defined) would be reasonably expected to
constitute a breaking change, and so we point our reusable actions to
FIREWHEEL major version `v2`. This prevents model component CI pipelines
from breaking in the event that FIREWHEEL bumps major versions and
substantially adjusts the installation process defined in the composite
actions in a way that has downstream effects on a given MC (which would
happen if the reusable actions referenced `main` instead of a major
version).