ci: Use additional files and reference tags for CI setup (#120)
This moves the anchors used for setup into a separate file
`.gitlab/ci/setup.yml` (and converts them to references). This allows
other GitLab repositories to include that YAML file directly instead of
including the main `.gitlab-ci.yml` file and then needing to override or
otherwise negate jobs that should not be run.

I also changed a couple of the job names and removed the `virtualenv`
dependency for Docker-based jobs in order to reuse a bunch of the
references in both the Docker-based jobs (docs, pages, build, deploy)
and the ones running in a native runner (performance). I didn't see an
obvious reason that we needed `virtualenv` beyond Python's built-in
`venv`, and it didn't play nicely with the native runner. If it is a
requirement, I'll find a workaround.