fix: Clearly delineate between editable/non-editable MC installs (#114)
This captures a handful of the higher priority elements introduced by
#91, while focusing on backwards compatibility.

The primary modification here is to make more clear the distinction
between cloned/uncloned editable/non-editable installs and have the docs
adequately reflect the actions in the `install.sh` script. Previously,
the documentation stated that the installer would clone our default
model components, which was true before we transitioned to the package
model. As we've made that transition, some of the implied functionality
has been lost—right now, running the install script with no arguments
will clone MCs but not install them... so still technically accurate,
but somewhat misleading because we do not mention that any further
actions to be taken. In a similar way, calling the script with
`--no-clone` will also not clone MCs, but will also not install them
even though one could argue that a familiar user (e.g., me) might expect
`--no-clone` to not _clone_ the MC repos, but still install them using
pip. There are a few other similar logic cases, but I won't discuss them
at length. Instead, I'll summarize the behavior of the script as would
be defined by this MR:

|  CLI Option          | (clone)  | `--no-clone` |
|------------------|---------|---------------|
| (production) | pip install model components from cloned versions | pip
install model components from PyPI |
| `--development` | pip install model components from cloned versions in
editable mode | pip install model components from PyPI (same as above) |

In all cases, when using this script, FIREWHEEL will be installed from
this local version not PyPI. I don't think that's actually the case
right now (looking at the line in the `install_firewheel` function), but
this seems like it would be the expected behavior of running this
script. Presumably, whoever's running the install script in the repo is
doing it because they want to install that version of FIREWHEEL, not
just have it turn around and download whatever the latest version is on
PyPI.