feat: Adding new docker build paramameters (#134)
This PR adds new build parameters that make it easy to modify the
creation of a FIREWHEEL docker container and corresponding documentation
to help override the default configuration options. Additionally, this
moves the docker container to use port 2222 by default as this will
prevent conflicts with host systems. This convention is used by other
virtualization systems, such as Vagrant.