fix: Fixing namespaces (#125)
This fixes FIREWHEEL working in a cluster. There are a few places where
the minimega namespace was hard-coded.

I cannot figure out why this change is necessary now, but everything
seemed to work in the past. Doesn't seem like the minimega API changed
at all. Not sure it's necessarily important, but would love to know if
anybody has comments.

TODO: 

- [x] Check if minimega API version matters and fix dependency
accordingly

---------

Signed-off-by: gregjacobus <30842019+gregjacobus@users.noreply.github.com>
Co-authored-by: Mitch Negus <21086604+mitchnegus@users.noreply.github.com>