fix: Refactor the MCPI to deduplicate MCs by absolute path (#141)
This MR refactors the `ModelComponentPathIterator` to rely on a
set-based backend that deduplicates MCs as it iterates.

## Background

This solves the duplicate MC errors that were previously seen when a
Python package based MC has source code in a repository that has also
been installed via `firewheel repository install ...`.

## Method

This MR gives the `ModelComponentPathIterator` the ability to remember
all the paths to all the MCs it has seen (and thus deduplicate them
using a `set`). However, because of how we loop over repos in the
`firewheel mc list` helper, this MR **does not** fix the fact that
duplicate repositories show in that display. That helper creates a new
path iterator for each repo, and so they do not know of MCs in other
repos. There are probably much more elegant dict-like data structures
that could be used to solve this issue, but I wanted to propose the MR
first before building anything along those lines.

### Impacts

The current set of changes leaves `ModelComponentPathIterator` largely
unchanged from the perspective of a user. (I believe the only change
from a user's perspective is that the return values of the
`walk_repository_for_model_component_paths` is a generator, not a list.)

### Extra Notes

1. My perspective is that having this `ModelComponentPathIterator` (and
the similar `ModelComponentIterator`) objects as an iterator is
unnecessarily complicated and provides minimal value. These objects
could be dramatically simplified by being _iterables_ (not _iterators_).
They could still evaluate lazily, which I think was probably the
original intention, they just wouldn't implement the `next(...)`
behavior. We don't use it very often, and it seems like it'd be more
straightforward for us to just loop over these objects like any other
iterable rather than call `next(...)`. Either way, this MR doesn't do
any work towards that goal.
2. I would also suggest that we remove the `repository_iterator`
attribute, but this is a public attribute so deleting it _could_ be a
breaking change. In our codebase(s) it is only ever used internal to the
`ModelComponentPathIterator` and this MR removes it from being used even
there.