chore: Remove code for building wheel in favor of CI deployments (#165)
This removes the call to build a FIREWHEEL package source distribution
from the install script. This should no longer be necessary now that
FIREWHEEL is either installed directly from PyPI or via an editable
install.

For applications where building a source distribution is necessary
(e.g., publishing to PyPI), that process should be handled
elsewhere—such as in CI job or in a local build workflow.