fix: Address issues where minimega environment variables are not passed (#158)
This PR addresses an issue that is most prevalent in Docker
environments, but could impact other experiments as well. We enable
users to overwrite minimega configuration values via environment
variables. However, our `restart` commands are run via SSH on all nodes
(via clustershell). Therefore, these variables must be passed to that
remote command much in the same way that we pass the python path and
GRPC path. This PR addresses resolves this issue.

Additionally, this corrects the larger problem of the environment
variables not being used properly within the docker environment. That
is, user-defined environment variables should be kept even if there are
other default values. This fix addresses the issue and ensures that our
container adheres to our docker documentation.

---------

Signed-off-by: Steven Elliott <6461548+sdelliot@users.noreply.github.com>
Co-authored-by: Mitch Negus <21086604+mitchnegus@users.noreply.github.com>