ci: Add flexibility to GitLab CI (#186)
# Add flexibility to GitLab CI

## Description
This adds flexibility to the GitLab CI script.

First, it adds a new rule to the performance testing, allowing a GitLab
CI/CD environment variable to be used for determining whether or not
performance tests should be run. The `FIREHWEEL_TEST_PERFORMANCE`
variable can be set to `true` for GitLab instances where performance
tests should be run. It can be set to `false` for GitLab instances where
performance tests are not desired.

Second, it uses a GitLab CI/CD variable for defining the tags of GitLab
runners that should run performance tests. This was previously hardcoded
`ubuntu2204`, but it could conceivably be named differently.


## Type of Change
Please select the type of change your pull request introduces:
- [ ] Bugfix
- [ ] New feature
- [ ] Documentation update
- [x] Other (please describe): CI/CD improvement

## Checklist
- [x] This PR conforms to the process detailed in the [Contributing
Guide](https://sandialabs.github.io/firewheel/developer/contributing.html).
- [x] I have included no proprietary/sensitive information in my code. 
- [x] I have performed a self-review of my code.
- [ ] _NA_ ~I have commented my code, particularly in hard-to-understand
areas.~
- [ ] _NA_ ~I have made corresponding changes to the documentation.~
- [x] My changes generate no new warnings.
- [x] I have tested my code.